wired up functions so things will load properly
