+++ api/src/openapi/schemas/survey.ts
@@ -1,0 +1,41 @@
+import { GeoJSONFeature } from './geoJson';
+
+export const SurveyLocationRequestObject = {
+  description: 'Survey location data',
+  type: 'array',
+  items: {
+    type: 'object',
+    required: [],
+    properties: {
+      survey_location_id: {
+        type: 'integer',
+        minimum: 1
+      },
+      name: {
+        type: 'string',
+        maxLength: 100
+      },
+      description: {
+        type: 'string',
+        maxLength: 250
+      },
+      geometry: {
+        type: 'string',
+        nullable: true
+      },
+      geography: {
+        type: 'string'
+      },
+      geojson: {
+        type: 'array',
+        items: {
+          ...(GeoJSONFeature as object)
+        }
+      },
+      revision_count: {
+        type: 'integer',
+        minimum: 0
+      }
+    }
+  }
+};
