+++ api/src/paths/project/{projectId}/survey/list.ts
@@ -3,7 +3,7 @@ import { Operation } from 'express-openapi';
 import { PROJECT_PERMISSION, SYSTEM_ROLE } from '../../../../constants/roles';
 import { getDBConnection } from '../../../../database/db';
 import { HTTP400 } from '../../../../errors/http-error';
-import { GeoJSONFeature } from '../../../../openapi/schemas/geoJson';
+import { SurveyLocationRequestObject } from '../../../../openapi/schemas/survey';
 import { authorizeRequestHandler } from '../../../../request-handlers/security/authorization';
 import { SurveyService } from '../../../../services/survey-service';
 import { getLogger } from '../../../../utils/logger';
@@ -254,37 +254,7 @@ GET.apiDoc = {
                       }
                     },
                     locations: {
-                      type: 'array',
-                      items: {
-                        type: 'object',
-                        required: ['survey_spatial_component_id', 'name', 'description', 'geometry', 'revision_count'],
-                        properties: {
-                          survey_spatial_component_id: {
-                            type: 'integer',
-                            minimum: 1,
-                            nullable: true
-                          },
-                          name: {
-                            type: 'string',
-                            maxLength: 100
-                          },
-                          description: {
-                            type: 'string',
-                            maxLength: 250
-                          },
-                          geometry: {
-                            type: 'array',
-                            items: {
-                              ...(GeoJSONFeature as object)
-                            }
-                          },
-                          revision_count: {
-                            type: 'integer',
-                            minimum: 0,
-                            nullable: true
-                          }
-                        }
-                      }
+                      ...(SurveyLocationRequestObject as object)
                     }
                   }
                 },
