+++ api/src/paths/project/{projectId}/survey/{surveyId}/update/get.ts
@@ -2,7 +2,7 @@ import { RequestHandler } from 'express';
 import { Operation } from 'express-openapi';
 import { PROJECT_PERMISSION, SYSTEM_ROLE } from '../../../../../../constants/roles';
 import { getDBConnection } from '../../../../../../database/db';
-import { GeoJSONFeature } from '../../../../../../openapi/schemas/geoJson';
+import { SurveyLocationRequestObject } from '../../../../../../openapi/schemas/survey';
 import { authorizeRequestHandler } from '../../../../../../request-handlers/security/authorization';
 import { SurveyService } from '../../../../../../services/survey-service';
 import { getLogger } from '../../../../../../utils/logger';
@@ -302,36 +302,7 @@ GET.apiDoc = {
                     }
                   },
                   locations: {
-                    description: 'Survey location data',
-                    type: 'array',
-                    items: {
-                      type: 'object',
-                      required: ['survey_spatial_component_id', 'name', 'description', 'geometry', 'revision_count'],
-                      properties: {
-                        survey_spatial_component_id: {
-                          type: 'integer',
-                          minimum: 1
-                        },
-                        name: {
-                          type: 'string',
-                          maxLength: 100
-                        },
-                        description: {
-                          type: 'string',
-                          maxLength: 250
-                        },
-                        geometry: {
-                          type: 'array',
-                          items: {
-                            ...(GeoJSONFeature as object)
-                          }
-                        },
-                        revision_count: {
-                          type: 'integer',
-                          minimum: 0
-                        }
-                      }
-                    }
+                    ...(SurveyLocationRequestObject as object)
                   },
                   participants: {
                     description: 'Survey participants Details',
