+++ api/src/services/survey-location-service.ts
@@ -1,6 +1,6 @@
 import { IDBConnection } from '../database/db';
 import { PostLocationData } from '../models/survey-create';
-import { SurveyLocationRepository } from '../repositories/survey-location-repository';
+import { SurveyLocationRecord, SurveyLocationRepository } from '../repositories/survey-location-repository';
 import { DBService } from './db-service';
 
 export class SurveyLocationService extends DBService {
@@ -14,5 +14,16 @@ export class SurveyLocationService extends DBService {
 
   async insertSurveyLocation(surveyId: number, data: PostLocationData): Promise<void> {
     return await this.surveyLocationRepository.insertSurveyLocation(surveyId, data);
+  }
+
+  /**
+   * Get Survey location for a given survey ID
+   *
+   * @param {number} surveyID
+   * @returns {*} {Promise<SurveyLocationRecord[]>}
+   * @memberof SurveyLocationService
+   */
+  async getSurveyLocationsData(surveyId: number): Promise<SurveyLocationRecord[]> {
+    return this.surveyLocationRepository.getSurveyLocationsData(surveyId);
   }
 }
